requirements.txt
click
requests
icalendar
dateutils